+++ src/main/java/xyz/mlhmz/gaspricelog/representation/dtos/ForecastGroupDto.java
@@ -2,6 +2,7 @@ package xyz.mlhmz.gaspricelog.representation.dtos;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -9,8 +10,8 @@ import java.util.UUID;
 public record ForecastGroupDto(
         UUID uuid,
         String groupName,
-        double gasPricePerKwh,
-        double kwhFactorPerQubicmeter,
+        BigDecimal gasPricePerKwh,
+        BigDecimal kwhFactorPerQubicmeter,
         List<EntryReferenceDto> entries,
         List<SpanReferenceDto> spans
 ) {
